requirements.txt
flask==2.0.1
pillow>=9.0.1
torch==2.0.1
torchvision==0.15.2
transformers==4.30.2
numpy==1.24.3
opencv-python==4.8.0.76
Werkzeug==2.3.7 